Move page container from pages to root
