+++ src/app.tsx
@@ -1,6 +1,6 @@
 import { HashRouter as Router, Switch, Route } from 'react-router-dom';
 import ChainDataContextProvider from './chain-data';
-import PageContainer from './pages/page-container';
+import PageContainer from './page-container';
 import Dashboard from './pages/dashboard';
 import Proposals from './pages/proposals';
 import ProposalDetails from './pages/proposal-commons/proposal-details';
